splitted Product-class functionality to DigitalProduct and PhysicalProduct extending BaseProduct and providing the necessary methods throught interface Product
