+++ src/main/java/com/github/mysterix5/DigitalProduct.java
@@ -1,0 +1,8 @@
+package com.github.mysterix5;
+
+public class DigitalProduct extends BaseProduct implements Product{
+
+    public DigitalProduct(String name){
+        super(name);
+    }
+}
