+++ src/main/java/com/github/mysterix5/PhysicalProduct.java
@@ -1,0 +1,8 @@
+package com.github.mysterix5;
+
+public class PhysicalProduct extends BaseProduct implements Product{
+
+    public PhysicalProduct(String name){
+        super(name);
+    }
+}
